Optimized all helper classes, removed most warnings

+++ src/main/java/net/ldm/mo_enchants/enchantment/helpers/LevitatingHelper.java
@@ -7,9 +7,6 @@ import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
-import net.minecraftforge.eventbus.api.Event;
-
-import javax.annotation.Nullable;
 
 public class LevitatingHelper {
 	public static void execute(Entity entity, Entity sourceentity) {
@@ -19,8 +16,8 @@ public class LevitatingHelper {
 				(sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)) != 0) {
 			if (entity instanceof LivingEntity _entity)
 				_entity.addEffect(new MobEffectInstance(MobEffects.LEVITATION,
-						(int) (EnchantmentHelper.getTagEnchantmentLevel(MoEnchantsEnchantments.LEVITATING.get(),
-								(sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)) * 10),
+						EnchantmentHelper.getTagEnchantmentLevel(MoEnchantsEnchantments.LEVITATING.get(),
+								(sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)) * 10,
 						0));
 		}
 	}
